+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   * Initial release.
 * SHSpaceUser
   * Initial release.
+* M365DSCUtil
+  * Added note about long paths being recommended when installing the Dev branch.
+    FIXES [#3158](https://github.com/microsoft/Microsoft365DSC/issues/3158)
 
 # 1.25.312.1