+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # UNRELEASED
 
+* AADServicePrincipal
+  * Removed logic that would translate AppId as a disply name from the
+    Set-TargetResource function.
 * IntuneDeviceConfigurationPolicyAndroidDeviceOwner
   * Fixed an issue where empty complex objects result in a broken export.
     FIXES [#5863](https://github.com/microsoft/Microsoft365DSC/issues/5863)
@@ -11,6 +14,10 @@
   * Initial release.
 * TeamsApplicationInstance
   * Initial release.
+* M365DSCDRGUtil
+  * Made a change in the Intune assignment evaluation logic not to validate that
+    the provided group is a valid Entra Id group based on GroupId. The logic
+    to validate this will be left to the Set-TargetResource.
 * M365DSCUtil
   * Added note about long paths being recommended when installing the Dev branch.
     FIXES [#3158](https://github.com/microsoft/Microsoft365DSC/issues/3158)